backend/requirements.txt
Flask
flask
Flask-SQLAlchemy
pymysql
flask-restplus
